% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_abstract.R
\name{AbstractComponent}
\alias{AbstractComponent}
\title{Abstract Interface for a Component}
\description{
AbstractComponent is the abstract base class for components.
All subclasses must implement fit and density.
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pi}}{the class probability of the component in the mixture model}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AbstractComponent-pi_density}{\code{AbstractComponent$pi_density()}}
\item \href{#method-AbstractComponent-density}{\code{AbstractComponent$density()}}
\item \href{#method-AbstractComponent-fit}{\code{AbstractComponent$fit()}}
\item \href{#method-AbstractComponent-set_pi}{\code{AbstractComponent$set_pi()}}
\item \href{#method-AbstractComponent-Q}{\code{AbstractComponent$Q()}}
\item \href{#method-AbstractComponent-to_fixed_component}{\code{AbstractComponent$to_fixed_component()}}
\item \href{#method-AbstractComponent-clone}{\code{AbstractComponent$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractComponent-pi_density"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractComponent-pi_density}{}}}
\subsection{Method \code{pi_density()}}{
Calculate the density multiplied with pi at the given x-coordinates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractComponent$pi_density(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{a vector of points at which the density is evaluated.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractComponent-density"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractComponent-density}{}}}
\subsection{Method \code{density()}}{
Abstract. Calculate the density at the given x-coordinates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractComponent$density(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{a vector of points at which the density is evaluated.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractComponent-fit"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractComponent-fit}{}}}
\subsection{Method \code{fit()}}{
Abstract. Fits the component to the signal rf by maximizing Q.
If performed approximately turns the algorithm into a ECM algorithm.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractComponent$fit(x, rf)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{coordinates of points on the x-axis.}

\item{\code{rf}}{coordinates of function values.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractComponent-set_pi"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractComponent-set_pi}{}}}
\subsection{Method \code{set_pi()}}{
Setter for pi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractComponent$set_pi(pi)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pi}}{the new value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractComponent-Q"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractComponent-Q}{}}}
\subsection{Method \code{Q()}}{
Calculates Q between rf and the density for the component as defined in !!!paper-doi!!!.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractComponent$Q(x, rf)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{vector of x values.}

\item{\code{rf}}{vector of the product between p_y and the responsiblity vector of the given component.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractComponent-to_fixed_component"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractComponent-to_fixed_component}{}}}
\subsection{Method \code{to_fixed_component()}}{
Returns a component with a fixed density identical to the density of the current component.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractComponent$to_fixed_component(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{vector of x values.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractComponent-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractComponent-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractComponent$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
