% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_abstract.R
\name{AbstractStrategy}
\alias{AbstractStrategy}
\title{Abstract Interface for a Peak Placement Strategy.}
\description{
AbstractStrategy is the abstract base class peak placement for strategies.
All subclasses must implement place_new_peak.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AbstractStrategy-place_new_peak}{\code{AbstractStrategy$place_new_peak()}}
\item \href{#method-AbstractStrategy-clone}{\code{AbstractStrategy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractStrategy-place_new_peak"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractStrategy-place_new_peak}{}}}
\subsection{Method \code{place_new_peak()}}{
Abstract. Calculate and return a new component.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractStrategy$place_new_peak(x, p_y, components)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{a vector of x coordinates.}

\item{\code{p_y}}{density being fit.}

\item{\code{components}}{list of AbstractComponent.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
new components list with the new peak added.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractStrategy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractStrategy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractStrategy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
