% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_fixed_density.R
\name{ComponentFixedDensity}
\alias{ComponentFixedDensity}
\title{Component which models a fixed background function.}
\description{
Requires fixed background function.
}
\details{
Component which models a fixed background function.
}
\keyword{internal}
\section{Super class}{
\code{\link[spectralem:AbstractComponent]{spectralem::AbstractComponent}} -> \code{ComponentFixedDensity}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{f}}{vector that defines density up to normalization.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ComponentFixedDensity-new}{\code{ComponentFixedDensity$new()}}
\item \href{#method-ComponentFixedDensity-density}{\code{ComponentFixedDensity$density()}}
\item \href{#method-ComponentFixedDensity-fit}{\code{ComponentFixedDensity$fit()}}
\item \href{#method-ComponentFixedDensity-get_amp}{\code{ComponentFixedDensity$get_amp()}}
\item \href{#method-ComponentFixedDensity-clone}{\code{ComponentFixedDensity$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="spectralem" data-topic="AbstractComponent" data-id="Q"><a href='../../spectralem/html/AbstractComponent.html#method-AbstractComponent-Q'><code>spectralem::AbstractComponent$Q()</code></a></span></li>
<li><span class="pkg-link" data-pkg="spectralem" data-topic="AbstractComponent" data-id="pi_density"><a href='../../spectralem/html/AbstractComponent.html#method-AbstractComponent-pi_density'><code>spectralem::AbstractComponent$pi_density()</code></a></span></li>
<li><span class="pkg-link" data-pkg="spectralem" data-topic="AbstractComponent" data-id="set_pi"><a href='../../spectralem/html/AbstractComponent.html#method-AbstractComponent-set_pi'><code>spectralem::AbstractComponent$set_pi()</code></a></span></li>
<li><span class="pkg-link" data-pkg="spectralem" data-topic="AbstractComponent" data-id="to_fixed_component"><a href='../../spectralem/html/AbstractComponent.html#method-AbstractComponent-to_fixed_component'><code>spectralem::AbstractComponent$to_fixed_component()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComponentFixedDensity-new"></a>}}
\if{latex}{\out{\hypertarget{method-ComponentFixedDensity-new}{}}}
\subsection{Method \code{new()}}{
Initializes a component with constant density proportional to f. Assumes f positive.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComponentFixedDensity$new(f)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{f}}{a positive vector.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComponentFixedDensity-density"></a>}}
\if{latex}{\out{\hypertarget{method-ComponentFixedDensity-density}{}}}
\subsection{Method \code{density()}}{
Calculate the density at the given x-coordinates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComponentFixedDensity$density(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{a vector of points at which the density is evaluated.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComponentFixedDensity-fit"></a>}}
\if{latex}{\out{\hypertarget{method-ComponentFixedDensity-fit}{}}}
\subsection{Method \code{fit()}}{
Fits the component to the signal rf.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComponentFixedDensity$fit(x, rf)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{coordinates of points on the x-axis.}

\item{\code{rf}}{coordinates of function values.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComponentFixedDensity-get_amp"></a>}}
\if{latex}{\out{\hypertarget{method-ComponentFixedDensity-get_amp}{}}}
\subsection{Method \code{get_amp()}}{
Gets amplitude relative to f.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComponentFixedDensity$get_amp()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComponentFixedDensity-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ComponentFixedDensity-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComponentFixedDensity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
