% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_trunct_linear.R
\name{ComponentTrunctLinear}
\alias{ComponentTrunctLinear}
\title{Component which models a linear background}
\description{
Component which models a linear background a x + b.
}
\section{Super class}{
\code{\link[spectralem:AbstractComponent]{spectralem::AbstractComponent}} -> \code{ComponentTrunctLinear}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{a}}{slope}

\item{\code{b}}{offset}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ComponentTrunctLinear-new}{\code{ComponentTrunctLinear$new()}}
\item \href{#method-ComponentTrunctLinear-density}{\code{ComponentTrunctLinear$density()}}
\item \href{#method-ComponentTrunctLinear-fit}{\code{ComponentTrunctLinear$fit()}}
\item \href{#method-ComponentTrunctLinear-set_pi}{\code{ComponentTrunctLinear$set_pi()}}
\item \href{#method-ComponentTrunctLinear-clone}{\code{ComponentTrunctLinear$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="spectralem" data-topic="AbstractComponent" data-id="Q"><a href='../../spectralem/html/AbstractComponent.html#method-AbstractComponent-Q'><code>spectralem::AbstractComponent$Q()</code></a></span></li>
<li><span class="pkg-link" data-pkg="spectralem" data-topic="AbstractComponent" data-id="pi_density"><a href='../../spectralem/html/AbstractComponent.html#method-AbstractComponent-pi_density'><code>spectralem::AbstractComponent$pi_density()</code></a></span></li>
<li><span class="pkg-link" data-pkg="spectralem" data-topic="AbstractComponent" data-id="to_fixed_component"><a href='../../spectralem/html/AbstractComponent.html#method-AbstractComponent-to_fixed_component'><code>spectralem::AbstractComponent$to_fixed_component()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComponentTrunctLinear-new"></a>}}
\if{latex}{\out{\hypertarget{method-ComponentTrunctLinear-new}{}}}
\subsection{Method \code{new()}}{
Initializes the linear component.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComponentTrunctLinear$new(pi = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pi}}{initial value for pi.}

\item{\code{nu_max}}{internal parameter, should be left as is.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComponentTrunctLinear-density"></a>}}
\if{latex}{\out{\hypertarget{method-ComponentTrunctLinear-density}{}}}
\subsection{Method \code{density()}}{
Calculate the density at the given x-coordinates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComponentTrunctLinear$density(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{a vector of points at which the density is evaluated.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComponentTrunctLinear-fit"></a>}}
\if{latex}{\out{\hypertarget{method-ComponentTrunctLinear-fit}{}}}
\subsection{Method \code{fit()}}{
Fits the component to the signal rf.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComponentTrunctLinear$fit(x, rf)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{coordinates of points on the x-axis.}

\item{\code{rf}}{coordinates of function values.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComponentTrunctLinear-set_pi"></a>}}
\if{latex}{\out{\hypertarget{method-ComponentTrunctLinear-set_pi}{}}}
\subsection{Method \code{set_pi()}}{
Setter for pi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComponentTrunctLinear$set_pi(pi)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pi}}{the new value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComponentTrunctLinear-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ComponentTrunctLinear-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComponentTrunctLinear$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
