% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_trunct_voigt.R
\name{ComponentTrunctVoigt}
\alias{ComponentTrunctVoigt}
\title{Component which models a truncated voigt profile}
\description{
Component which models a truncated voigt profile. Models a voigt peak in the signal.
Truncation is needed to deal with peaks which have significant mass outside the x range.
}
\keyword{internal}
\section{Super class}{
\code{\link[spectralem:AbstractComponent]{spectralem::AbstractComponent}} -> \code{ComponentTrunctVoigt}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pos}}{position of the voigt profile}

\item{\code{gwidth}}{gaussian width of the voigt profile}

\item{\code{lwidth}}{gaussian width of the voigt profile}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ComponentTrunctVoigt-new}{\code{ComponentTrunctVoigt$new()}}
\item \href{#method-ComponentTrunctVoigt-density}{\code{ComponentTrunctVoigt$density()}}
\item \href{#method-ComponentTrunctVoigt-fit}{\code{ComponentTrunctVoigt$fit()}}
\item \href{#method-ComponentTrunctVoigt-set_params}{\code{ComponentTrunctVoigt$set_params()}}
\item \href{#method-ComponentTrunctVoigt-get_amp}{\code{ComponentTrunctVoigt$get_amp()}}
\item \href{#method-ComponentTrunctVoigt-clone}{\code{ComponentTrunctVoigt$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="spectralem" data-topic="AbstractComponent" data-id="Q"><a href='../../spectralem/html/AbstractComponent.html#method-AbstractComponent-Q'><code>spectralem::AbstractComponent$Q()</code></a></span></li>
<li><span class="pkg-link" data-pkg="spectralem" data-topic="AbstractComponent" data-id="pi_density"><a href='../../spectralem/html/AbstractComponent.html#method-AbstractComponent-pi_density'><code>spectralem::AbstractComponent$pi_density()</code></a></span></li>
<li><span class="pkg-link" data-pkg="spectralem" data-topic="AbstractComponent" data-id="set_pi"><a href='../../spectralem/html/AbstractComponent.html#method-AbstractComponent-set_pi'><code>spectralem::AbstractComponent$set_pi()</code></a></span></li>
<li><span class="pkg-link" data-pkg="spectralem" data-topic="AbstractComponent" data-id="to_fixed_component"><a href='../../spectralem/html/AbstractComponent.html#method-AbstractComponent-to_fixed_component'><code>spectralem::AbstractComponent$to_fixed_component()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComponentTrunctVoigt-new"></a>}}
\if{latex}{\out{\hypertarget{method-ComponentTrunctVoigt-new}{}}}
\subsection{Method \code{new()}}{
Initialize the voigt component.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComponentTrunctVoigt$new(min_width, possible_peak_positions, pi = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{min_width}}{minimal value for lwidth and gwidth}

\item{\code{possible_peak_positions}}{range of possible peak positions}

\item{\code{pi}}{initial value for pi.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComponentTrunctVoigt-density"></a>}}
\if{latex}{\out{\hypertarget{method-ComponentTrunctVoigt-density}{}}}
\subsection{Method \code{density()}}{
Calculate the density at the given x-coordinates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComponentTrunctVoigt$density(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{a vector of points at which the density is evaluated.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComponentTrunctVoigt-fit"></a>}}
\if{latex}{\out{\hypertarget{method-ComponentTrunctVoigt-fit}{}}}
\subsection{Method \code{fit()}}{
Fits the component to the signal rf by maximizing Q.
Ensures monotonicity by comparing with last fit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComponentTrunctVoigt$fit(x, rf)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{coordinates of points on the x-axis.}

\item{\code{rf}}{coordinates of function values.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComponentTrunctVoigt-set_params"></a>}}
\if{latex}{\out{\hypertarget{method-ComponentTrunctVoigt-set_params}{}}}
\subsection{Method \code{set_params()}}{
Setter for pos, gwidth and lwidth.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComponentTrunctVoigt$set_params(pos = NULL, gwidth = NULL, lwidth = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pos}}{the new pos}

\item{\code{gwidth}}{the new gwidth}

\item{\code{lwidth}}{the new lwidth}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComponentTrunctVoigt-get_amp"></a>}}
\if{latex}{\out{\hypertarget{method-ComponentTrunctVoigt-get_amp}{}}}
\subsection{Method \code{get_amp()}}{
Gets amplitude of the voigt profile.

Calculate Q and it's gradient.

Guess pos, gwidth, lwidth by fitting a pseudo voigt. To be treated as a blackbox.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComponentTrunctVoigt$get_amp()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComponentTrunctVoigt-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ComponentTrunctVoigt-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComponentTrunctVoigt$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
