% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{data.synthetic.stormyclouds}
\alias{data.synthetic.stormyclouds}
\title{A synthethic Dataset.}
\usage{
data.synthetic.stormyclouds(seed, kp, noise)
}
\arguments{
\item{seed}{sets the random seed.}

\item{kp}{number of peaks.}

\item{noise}{sigma of gaussian noise.}
}
\description{
Return list of x and y and the true parameters.
}
