% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{integrate}
\alias{integrate}
\title{Compute the area of a function with values 'y' at the points 'x'.}
\usage{
integrate(x, y)
}
\arguments{
\item{x}{coordinates of points on the x-axis}

\item{y}{coordinates of function values}
}
\description{
Wrapper function for trapezoidal integration from pracma.
}
