% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_max_smooth_lossfield.R
\name{StrategyMaxSmoothLossfield}
\alias{StrategyMaxSmoothLossfield}
\title{Max Smoothed Lossfield Placement Strategy}
\description{
Places a new Voigt peak where the lossfield is maximal.
}
\details{
Max Smoothed Lossfield Placement Strategy
}
\keyword{internal}
\section{Super class}{
\code{\link[spectralem:AbstractStrategy]{spectralem::AbstractStrategy}} -> \code{StrategyMaxSmoothLossfield}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{typical_width}}{typical width of peaks, used as the width of the blur kernel and for the gwidth of the newly placed peak}

\item{\code{min_width}}{minimal width of a placed peak}

\item{\code{burn_in_iters}}{number of iterations to perform with fixed densities}

\item{\code{possible_peak_positions}}{range of possible peak positions}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StrategyMaxSmoothLossfield-new}{\code{StrategyMaxSmoothLossfield$new()}}
\item \href{#method-StrategyMaxSmoothLossfield-place_new_peak}{\code{StrategyMaxSmoothLossfield$place_new_peak()}}
\item \href{#method-StrategyMaxSmoothLossfield-clone}{\code{StrategyMaxSmoothLossfield$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StrategyMaxSmoothLossfield-new"></a>}}
\if{latex}{\out{\hypertarget{method-StrategyMaxSmoothLossfield-new}{}}}
\subsection{Method \code{new()}}{
Initialize the strategy.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrategyMaxSmoothLossfield$new(
  min_width,
  typical_width,
  burn_in_iters,
  possible_peak_positions
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{min_width}}{minimal gwidth and lwidth for the newly placed peak}

\item{\code{typical_width}}{typical width of peaks, used as the width of the blur kernel and for the gwidth of the newly placed peak}

\item{\code{burn_in_iters}}{number of iterations to perform with fixed densities.}

\item{\code{possible_peak_positions}}{range of possible peak positions}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StrategyMaxSmoothLossfield-place_new_peak"></a>}}
\if{latex}{\out{\hypertarget{method-StrategyMaxSmoothLossfield-place_new_peak}{}}}
\subsection{Method \code{place_new_peak()}}{
Calculate and return a new component (usually a voigt component).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrategyMaxSmoothLossfield$place_new_peak(x, p_y, components)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{a vector of x coordinates.}

\item{\code{p_y}}{density being fit.}

\item{\code{components}}{a list of AbstractComponents.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
new components list with the new peak added.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StrategyMaxSmoothLossfield-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StrategyMaxSmoothLossfield-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrategyMaxSmoothLossfield$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
