% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{synthetic.signal}
\alias{synthetic.signal}
\title{A synthethic Raman signal}
\usage{
synthetic.signal(seed, K, noise)
}
\arguments{
\item{seed}{sets the random seed}

\item{K}{number of peaks}

\item{noise}{std of noise}
}
\description{
Returns list of `x` and `y` and the true parameters.
}
