% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{voigt.model}
\alias{voigt.model}
\title{Signal model for Voigt profiles}
\usage{
voigt.model(x, params)
}
\arguments{
\item{x}{vector of \code{x} coordinates the model is to be evaluated at}

\item{params}{parameter list as described in the description}
}
\description{
Implements the signal model discussed in `spectralem()`. Doesn't include any \eqn{f_i}.

Parameters are passed in a list params with components \code{amp}, \code{pos}, \code{gwidth} and \code{lwidth} which are vectors of the same length.
For each the i-th entry yields the parameters of the i-th Voigt profile.
Furthermore named components \code{a} and \code{b} describe the slope and offset of the linear background respectively. See examples.
}
\examples{
y <- voigt.model(
  x = seq(0, 1000),
  params = list(
    amp = c(10, 2, 3, 3),
    pos = c(332, 577, 697),
    gwidth = c(5, 5, 5),
    lwidth = c(0.1, 0.1, 0.1),
    a = 0.01,
    b = 15
  )
)
}
