% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm.R
\name{spectralem}
\alias{spectralem}
\title{Quickly fits voigt profiles to a signal.}
\usage{
spectralem(
  x,
  y,
  max_peaks,
  max_iter,
  print_progress = TRUE,
  add_component_every_iters = 10,
  start_peaks = list(pos = c(), gwidth = c(), lwidth = c()),
  background_model = list(ComponentTrunctLinear$new()),
  placement_strategy = StrategyMaxSmoothLossfield$new()
)
}
\description{
Requires no starting fit.
Most useful to get a fast and decent fit.
The algorithm was developed to produce a fast and decent fit that general non linear optimizers such as optim fail to produce.
}
